feat: 정기 배송 api 구현

+++ src/main/java/com/petmily/backend/domain/order/entity/OrderItem.java
@@ -54,7 +54,7 @@ public class OrderItem extends BaseTimeEntity {
         return OrderItem.builder()
             .productId(product.getId())
             .quantity(quantity)
-            .price(product.getPrice())
+            .price(BigDecimal.valueOf(product.getPrice()))
             .build();
     }
